+++ app/router.ts
@@ -8,10 +8,13 @@ export default class Router extends EmberRouter {
 
 Router.map(function () {
   // Add route declarations here
-  this.route('login');
   this.route('about');
   this.route('applog');
   this.route('changelog');
+  this.route('auth', function () {
+    this.route('login');
+    this.route('callback');
+  });
   this.route('authenticated', { path: '/' }, function () {
     this.route('home');
     this.route('board-overview');